LayoutTests/imported/w3c/web-platform-tests/geolocation-API/disabled-by-permissions-policy.https.sub-expected.txt

CONSOLE MESSAGE: Permission policy 'Geolocation' check failed for element with origin 'https://127.0.0.1:9443' and allow attribute ''.


FAIL Permissions-Policy header geolocation=() disallows the top-level document. promise_test: Unhandled rejection with value: object "[object GeolocationPosition]"
FAIL Permissions-Policy header geolocation=() disallows same-origin iframes. assert_false: geolocation expected false got true
PASS Permissions-Policy header geolocation=() disallows cross-origin iframes.
